more clangtidy

+++ offline/framework/fun4allraw/SingleMicromegasPoolInput_v2.cc
@@ -667,12 +667,12 @@ void SingleMicromegasPoolInput_v2::process_fee_data( int packet_id, unsigned int
     // number of 10-bit words in this packet
     payload.adc_length = data_buffer[0] - HEADER_LENGTH;
     payload.data_parity = data_buffer[4] >> 9U;
-    payload.sampa_address = (data_buffer[4] >> 5U) & 0xfU;
+    payload.sampa_address = (uint16_t)(data_buffer[4] >> 5U) & 0xfU;
     payload.sampa_channel = data_buffer[4] & 0x1fU;
     payload.channel = data_buffer[4] & 0x1ffU;
-    payload.type = (data_buffer[3] >> 7U) & 0x7U;
+    payload.type = (uint16_t)(data_buffer[3] >> 7U) & 0x7U;
     payload.user_word = data_buffer[3] & 0x7fU;
-    payload.bx_timestamp = ((data_buffer[6] & 0x3ffU) << 10U) | (data_buffer[5] & 0x3ffU);
+    payload.bx_timestamp = ((uint16_t)(data_buffer[6] & 0x3ffU) << 10U) | (uint16_t)(data_buffer[5] & 0x3ffU);
 
     // crc
     payload.data_crc = data_buffer[pkt_length];
